% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{acceptance.rate.plot}
\alias{acceptance.rate.plot}
\title{Acceptance rate plot}
\usage{
acceptance.rate.plot(accepted, burn.in = 0)
}
\arguments{
\item{accepted}{vector of binary variable for acceptances (TRUE or 1=accepted,
FALSE or 0=rejected)}

\item{burn.in}{burn-in draws to drop}
}
\description{
Acceptance rate plot
}
