% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfunc-package.r
\docType{package}
\name{acfunc}
\alias{acfunc}
\alias{acfunc-package}
\title{acfunc - Alex Cooper's useful functions}
\description{
This package includes statistical functions I have found useful.
You are welcome to use it, but you do so at your own risk.
}
