% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdist-package.r
\docType{package}
\name{bdist}
\alias{bdist}
\alias{bdist-package}
\title{bdist - Extra Distributions for Bayesian Analysis}
\description{
This package includes statistical functions I have found useful.
You are welcome to use it, but you do so at your own risk.
}
