% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse-gamma.R
\name{dinvgamma}
\alias{dinvgamma}
\alias{rinvgamma}
\alias{pinvgamma}
\title{Inverse gamma distribution}
\usage{
dinvgamma(x, shape, scale, log = FALSE)

rinvgamma(n, shape, scale)

pinvgamma(q, shape, scale, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{shape}{shape parameter alpha; must be positive}

\item{scale}{scale parameter beta; must be positive}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of random deviates to draw}

\item{q}{vector of quantiles}
}
\value{
`dinvgamma` gives the density and `rinvgamma` generates random deviates.
}
\description{
We follow Gelman et al's (2014) parameterization. If \eqn{X\sim Gamma(a,b)}, then \eqn{1/X\sim InvGamma(a,b)}.
Note that some authors use alternative parameterizations; see especially \link{dinvrootgamma}.
}
\details{
The density function with shape \eqn{\alpha} and rate \eqn{\beta} is
\deqn{\frac{\beta ^{\alpha } x^{-\alpha -1} \exp\{-\frac{\beta }{x}\}}{\Gamma (\alpha )}.}

The cdf with shape \eqn{\alpha} and scale \eqn{\beta} is
\deqn{\frac{\Gamma \left(\alpha ,\frac{\beta }{x}\right)}{\Gamma (\alpha )}.}
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., & Rubin, D. B. (2014). Bayesian data analysis (3E). Boca Raton, FL, USA: Chapman & Hall/CRC.
}
\seealso{
Other gamma: \code{\link{dinvrootgamma}}
}
