% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse-gamma.R
\name{dinvrootgamma}
\alias{dinvrootgamma}
\alias{rinvrootgamma}
\alias{pinvrootgamma}
\title{Inverse root gamma distribution}
\usage{
dinvrootgamma(x, shape, df, log = FALSE)

rinvrootgamma(n, shape, df)

pinvrootgamma(q, shape, df, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{shape}{parameter, where shape>0}

\item{df}{degrees of freedom parameter, where df>0}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of random deviates to draw}

\item{q}{vector of quantiles}
}
\value{
`dinvgamma` gives the density and `rinvgamma` generates random deviates.
}
\description{
If X ~ InvRootGamma(scale=sigma.sq, df=nu), then
1/(X^2) ~ Gamma(shape=nu/2, rate=nu*sigma.sq/2).
}
\details{
The density function of the inverse-root-gamma distribution shape \eqn{\sigma^2}
and \eqn{\nu} degrees of freedom is
\deqn{p(\theta) = I(\theta>0)\frac{2}{\Gamma\left(\frac{\nu}{2}\right)}
      \left( \frac{\nu\sigma^2}{2}\right)^{\nu/2}\frac{1}{\theta^{\nu+1}}
      \exp\left\{-\frac{\nu \sigma^2}{2\theta^2}\right\}.}

The cdf is
\deqn{P(x) = \frac{\Gamma \left(\frac{\nu }{2},\frac{\nu  \sigma ^2}{2 x^2}\right)}{\Gamma \left(\frac{\nu }{2}\right)}.}

Note that some authors use alternative parameterizations; see especially \link{dinvgamma}.
}
\seealso{
Other gamma: \code{\link{dinvgamma}}
}
