% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{dnorm_trunc}
\alias{dnorm_trunc}
\title{Density of a truncated normal distribution}
\usage{
dnorm_trunc(x, lower = -Inf, upper = Inf, mean = 0, sd = 1, log = F)
}
\arguments{
\item{x}{quantile at which to evaluate density}

\item{lower}{lower truncation bound}

\item{upper}{upper truncation bound}

\item{mean}{underlying distribution mean}

\item{sd}{underlying distribution sd}

\item{log}{if true, return log density}
}
\description{
Density of a truncated normal distribution
}
