% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student.R
\name{dst}
\alias{dst}
\alias{rst}
\alias{pst}
\title{Student t distribution with location mu, scale sigma, and nu degrees of freedom.}
\usage{
dst(x, df, location = 0, scale = 1, log = FALSE)

rst(n, df, location = 0, scale = 1)

pst(q, df, location = 0, scale = 1)
}
\arguments{
\item{x}{vector of quantiles}

\item{df}{degrees of freedom}

\item{location}{mu parameter}

\item{scale}{parameter (can be written sigma or sigma^2; this is 'sigma^2' in the above expression)}

\item{log}{return logarithm of value if true}

\item{n}{number of random deviates to draw}

\item{q}{vector of quantiles}
}
\value{
`dst` gives the density and `rst` generates random deviates.
}
\description{
The pdf of the Student t distribution is given in Gelman et al (2014, p.578):
\deqn{t(x | \mu, \sigma, \nu) = \frac{\gamma\left(\frac{\nu+1}{2}\right)}{\gamma(\frac{\nu}{2})\sqrt{\nu\pi}\sigma}\left(1 + \frac{1}{\nu}\left(\frac{x-\mu}{\sigma}\right)^2\right)^{-\frac{\nu+1}{2}}}
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., & Rubin, D. B. (2014). Bayesian data analysis (3E). Boca Raton, FL, USA: Chapman & Hall/CRC.
}
