% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp.R
\name{dgdp}
\alias{dgdp}
\alias{rgdp}
\alias{pgdp}
\alias{qgdp}
\title{Generalized double Pareto distribution.}
\usage{
dgdp(x, xi, alpha)

rgdp(n, xi, alpha)

pgdp(q, xi, alpha)

qgdp(p, xi, alpha)
}
\arguments{
\item{x}{vector of quantiles}

\item{xi}{scale parameter, where xi > 0}

\item{alpha}{shape parameter, where alpha > 0}

\item{n}{number of variates to simulate}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}
}
\description{
If X ~ gdP(alpha, xi), then it has density
}
\details{
f(X|alpha,xi) = 1/(2*xi) * (1 + abs(X)/(alpha*xi))^(-alpha-1)
}
\references{
Armagan, Artin, David B. Dunson, and Jaeyong Lee. "Generalized
            double Pareto shrinkage." Statistica Sinica 23.1 (2013): 119.
}
