% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{marginal.plots}
\alias{marginal.plots}
\title{Marginal density plots, optionally discarding burn-in draws}
\usage{
marginal.plots(draws, burn.in = 0)
}
\arguments{
\item{draws}{data frame of draws, one column per variable}

\item{burn.in}{number of draws to drop (default 0)}
}
\description{
Marginal density plots, optionally discarding burn-in draws
}
