% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{mvdnorm}
\alias{mvdnorm}
\alias{mvrnorm}
\title{Multivariate normal distribution}
\usage{
mvdnorm(x, mu, Sigma, tol = 1e-06)

mvrnorm(n, mu, Sigma, tol = 1e-06)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{mean, a d-element vector}

\item{Sigma}{symmetric, positive definite d*d variance-covariance matrix}

\item{tol}{tolerance for checking positive definiteness}

\item{n}{number of variates to draw}
}
\description{
Multivariate normal distribution
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., & Rubin, D. B. (2014). Bayesian data analysis (3E). Boca Raton, FL, USA: Chapman & Hall/CRC.
}
