% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{rnorm_trunc}
\alias{rnorm_trunc}
\alias{dnorm_trunc}
\alias{pnorm_trunc}
\alias{qnorm_trunc}
\title{Truncated normal distribution.}
\usage{
rnorm_trunc(n, mean, sd, lower, upper)

dnorm_trunc(x, lower = -Inf, upper = Inf, mean = 0, sd = 1, log = F)

pnorm_trunc(q, mean, sd, lower, upper)

qnorm_trunc(p, mean, sd, lower, upper)
}
\arguments{
\item{n}{number of variates to generate}

\item{mean}{underlying distribution mean}

\item{sd}{underlying distribution sd}

\item{lower}{lower truncation bound}

\item{upper}{upper truncation bound}

\item{x}{vector of quantiles}

\item{log}{return logarithm}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}
}
\description{
Density, distribution function, quantile function, and random generation
for the normal distribution truncated at (lower, upper).
}
