% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.R
\name{rnorm1_trunc}
\alias{rnorm1_trunc}
\title{Draw 1 variate from truncated normal distribution.
Runs in Op(1) time and space. Stack can overrun if range too far from draws.}
\usage{
rnorm1_trunc(lower, upper, mean, sd)
}
\arguments{
\item{lower}{lower truncation bound}

\item{upper}{upper truncation bound}

\item{mean}{underlying distribution mean}

\item{sd}{underlying distribution sd}
}
\description{
Draw 1 variate from truncated normal distribution.
Runs in Op(1) time and space. Stack can overrun if range too far from draws.
}
