% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull.R
\name{rweibull2}
\alias{rweibull2}
\alias{dweibull2}
\alias{qweibull2}
\alias{pweibull2}
\title{Weibull distribution}
\usage{
rweibull2(n, alpha, gamma)

dweibull2(x, alpha, gamma)

qweibull2(q, alpha, gamma)

pweibull2(q, alpha, gamma)
}
\arguments{
\item{n}{number of variates to draw}

\item{alpha}{alpha parameter}

\item{gamma}{gamma parameter}

\item{q}{vector of quantiles}
}
\description{
Density, distribution function, quantile function, and random generation
for the weibull distribution. This is different from the standard R
weibull distribution, differing in its parameterization.
}
\details{
The density of a weibull-distributed variable X ~ W(alpha, beta) is
f(t) = gamma * alpha * t^(alpha-1) * exp(-gamma * t^alpha)
}
