% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_plots.R
\name{trace.plot}
\alias{trace.plot}
\title{Generate MCMC trace plots based on a data.frame of draws.}
\usage{
trace.plot(draws, burn.in = 0)
}
\arguments{
\item{draws}{data frame of draws, one column per variable}

\item{burn.in}{number of burn-in draws (default 0)}
}
\description{
Generate MCMC trace plots based on a data.frame of draws.
}
